Do not mv a file that does not exist

+++ Virtual/Machine/ISO-maker/Files/For-Target/install-xubuntu-desktop.sh
@@ -13,14 +13,16 @@ build_date="$(date +%Y%m%d%H%S)"
 backdrops=usr/share/xfce4/backdrops
 
 if [[ ! -L "/$backdrops/xubuntu-wallpaper.png"  ]]; then # original config
-    sudo mv /$backdrops/xubuntu-wallpaper.png         /$backdrops/xubuntu-wallpaper.png_$build_date
-    sudo ln -s /var/local/root/$backdrops/Econ-ARK-Logo-1536x768.png /$backdrops/xubuntu-wallpaper.png 
+    sudo mv /$backdrops/xubuntu-wallpaper.png         "/$backdrops/xubuntu-wallpaper.png_$build_date"
+    sudo cp  /var/local/root/$backdrops/Econ-ARK-Logo-1536x768.png /$backdrops/xubuntu-wallpaper.png 
 fi
 
 # Document, in /var/local, where its content is used
 ## Move but preserve the original
 sudo mv                /usr/share/lightdm/lightdm.conf.d/60-xubuntu.conf /usr/share/lightdm/lightdm.conf.d/60-xubuntu.conf_$build_date
-ln -s   /var/local/root/usr/share/lightdm/lightdm.conf.d/60-xubuntu.conf /usr/share/lightdm/lightdm.conf.d/60-xubuntu.conf
+if [[ -e /usr/share/lightdm/lightdm.conf.d/60-xubuntu.conf ]] && [[ -s /usr/share/lightdm/lightdm.conf.d/60-xubuntu.conf ]]; then
+    ln -s   /var/local/root/usr/share/lightdm/lightdm.conf.d/60-xubuntu.conf /usr/share/lightdm/lightdm.conf.d/60-xubuntu.conf
+fi
 ## Do not start ubuntu at all
 [[ -e /usr/share/lightdm/lightdm.conf.d/50-ubuntu.conf ]] && sudo mv     /usr/share/lightdm/lightdm.conf.d/50-ubuntu.conf /usr/share/lightdm/lightdm.conf.d/50-ubuntu.conf_$build_date
